+++ src/atlas/functionspace/NodeColumnsInterface.cc
@@ -246,7 +246,7 @@ void atlas__NodesFunctionSpace__sum_arr_double(const NodeColumns* This, const fi
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new double[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -261,7 +261,7 @@ void atlas__NodesFunctionSpace__sum_arr_float(const NodeColumns* This, const fie
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new float[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -276,7 +276,7 @@ void atlas__NodesFunctionSpace__sum_arr_long(const NodeColumns* This, const fiel
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new long[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -291,7 +291,7 @@ void atlas__NodesFunctionSpace__sum_arr_int(const NodeColumns* This, const field
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new int[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -325,7 +325,7 @@ void atlas__NodesFunctionSpace__oisum_arr_double(const NodeColumns* This, const 
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new double[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -340,7 +340,7 @@ void atlas__NodesFunctionSpace__oisum_arr_float(const NodeColumns* This, const f
         This->orderIndependentSum(*field,sumvec,size_t_N);
         size = sumvec.size();
         sum = new float[size];
-        for( size_t j=0; j<size; ++j ) sum[j] = sumvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) sum[j] = sumvec[j];
   );
   N = size_t_N;
 }
@@ -410,7 +410,7 @@ void atlas__NodesFunctionSpace__min_arr_double(const NodeColumns* This, const fi
         This->minimum(*field,minvec);
         size = minvec.size();
         minimum = new double[size];
-        for( size_t j=0; j<size; ++j ) minimum[j] = minvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) minimum[j] = minvec[j];
   );
 }
 
@@ -423,7 +423,7 @@ void atlas__NodesFunctionSpace__min_arr_float(const NodeColumns* This, const fie
         This->minimum(*field,minvec);
         size = minvec.size();
         minimum = new float[size];
-        for( size_t j=0; j<size; ++j ) minimum[j] = minvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) minimum[j] = minvec[j];
   );
 }
 
@@ -436,7 +436,7 @@ void atlas__NodesFunctionSpace__min_arr_long(const NodeColumns* This, const fiel
         This->minimum(*field,minvec);
         size = minvec.size();
         minimum = new long[size];
-        for( size_t j=0; j<size; ++j ) minimum[j] = minvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) minimum[j] = minvec[j];
   );
 }
 
@@ -449,7 +449,7 @@ void atlas__NodesFunctionSpace__min_arr_int(const NodeColumns* This, const field
         This->minimum(*field,minvec);
         size = minvec.size();
         minimum = new int[size];
-        for( size_t j=0; j<size; ++j ) minimum[j] = minvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) minimum[j] = minvec[j];
   );
 }
 
@@ -463,7 +463,7 @@ void atlas__NodesFunctionSpace__max_arr_double(const NodeColumns* This, const fi
         This->maximum(*field,maxvec);
         size = maxvec.size();
         maximum = new double[size];
-        for( size_t j=0; j<size; ++j ) maximum[j] = maxvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) maximum[j] = maxvec[j];
   );
 }
 
@@ -476,7 +476,7 @@ void atlas__NodesFunctionSpace__max_arr_float(const NodeColumns* This, const fie
         This->maximum(*field,maxvec);
         size = maxvec.size();
         maximum = new float[size];
-        for( size_t j=0; j<size; ++j ) maximum[j] = maxvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) maximum[j] = maxvec[j];
   );
 }
 
@@ -489,7 +489,7 @@ void atlas__NodesFunctionSpace__max_arr_long(const NodeColumns* This, const fiel
         This->maximum(*field,maxvec);
         size = maxvec.size();
         maximum = new long[size];
-        for( size_t j=0; j<size; ++j ) maximum[j] = maxvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) maximum[j] = maxvec[j];
   );
 }
 
@@ -502,7 +502,7 @@ void atlas__NodesFunctionSpace__max_arr_int(const NodeColumns* This, const field
         This->maximum(*field,maxvec);
         size = maxvec.size();
         maximum = new int[size];
-        for( size_t j=0; j<size; ++j ) maximum[j] = maxvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) maximum[j] = maxvec[j];
   );
 }
 
@@ -589,7 +589,7 @@ void atlas__NodesFunctionSpace__minloc_arr_double(const NodeColumns* This, const
         size = minvec.size();
         minimum = new double[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -607,7 +607,7 @@ void atlas__NodesFunctionSpace__minloc_arr_float(const NodeColumns* This, const 
         size = minvec.size();
         minimum = new float[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -625,7 +625,7 @@ void atlas__NodesFunctionSpace__minloc_arr_long(const NodeColumns* This, const f
         size = minvec.size();
         minimum = new long[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -643,7 +643,7 @@ void atlas__NodesFunctionSpace__minloc_arr_int(const NodeColumns* This, const fi
         size = minvec.size();
         minimum = new int[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -661,7 +661,7 @@ void atlas__NodesFunctionSpace__maxloc_arr_double(const NodeColumns* This, const
         size = maxvec.size();
         maximum = new double[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -679,7 +679,7 @@ void atlas__NodesFunctionSpace__maxloc_arr_float(const NodeColumns* This, const 
         size = maxvec.size();
         maximum = new float[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -697,7 +697,7 @@ void atlas__NodesFunctionSpace__maxloc_arr_long(const NodeColumns* This, const f
         size = maxvec.size();
         maximum = new long[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -715,7 +715,7 @@ void atlas__NodesFunctionSpace__maxloc_arr_int(const NodeColumns* This, const fi
         size = maxvec.size();
         maximum = new int[size];
         glb_idx = new long[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
         }
@@ -760,7 +760,7 @@ void atlas__NodesFunctionSpace__mean_arr_double(const NodeColumns* This, const f
         This->mean(*field,meanvec,size_t_N);
         size = meanvec.size();
         mean = new double[size];
-        for( size_t j=0; j<size; ++j ) mean[j] = meanvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) mean[j] = meanvec[j];
   );
   N = size_t_N;
 }
@@ -775,7 +775,7 @@ void atlas__NodesFunctionSpace__mean_arr_float(const NodeColumns* This, const fi
         This->mean(*field,meanvec,size_t_N);
         size = meanvec.size();
         mean = new float[size];
-        for( size_t j=0; j<size; ++j ) mean[j] = meanvec[j];
+        for( size_t j=0; j<(size_t)size; ++j ) mean[j] = meanvec[j];
   );
   N = size_t_N;
 }
@@ -830,7 +830,7 @@ void atlas__NodesFunctionSpace__mean_and_stddev_arr_double(const NodeColumns* Th
         size = meanvec.size();
         mean = new double[size];
         stddev = new double[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           mean[j] = meanvec[j];
           stddev[j] = stddevvec[j];
         }
@@ -851,7 +851,7 @@ void atlas__NodesFunctionSpace__mean_and_stddev_arr_float(const NodeColumns* Thi
         size = meanvec.size();
         mean = new float[size];
         stddev = new float[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           mean[j] = meanvec[j];
           stddev[j] = stddevvec[j];
         }
@@ -970,7 +970,7 @@ void atlas__NodesFunctionSpace__minloclev_arr_double(const NodeColumns* This, co
         minimum = new double[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -991,7 +991,7 @@ void atlas__NodesFunctionSpace__minloclev_arr_float(const NodeColumns* This, con
         minimum = new float[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1012,7 +1012,7 @@ void atlas__NodesFunctionSpace__minloclev_arr_long(const NodeColumns* This, cons
         minimum = new long[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1033,7 +1033,7 @@ void atlas__NodesFunctionSpace__minloclev_arr_int(const NodeColumns* This, const
         minimum = new int[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           minimum[j] = minvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1054,7 +1054,7 @@ void atlas__NodesFunctionSpace__maxloclev_arr_double(const NodeColumns* This, co
         maximum = new double[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1075,7 +1075,7 @@ void atlas__NodesFunctionSpace__maxloclev_arr_float(const NodeColumns* This, con
         maximum = new float[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1096,7 +1096,7 @@ void atlas__NodesFunctionSpace__maxloclev_arr_long(const NodeColumns* This, cons
         maximum = new long[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
@@ -1117,7 +1117,7 @@ void atlas__NodesFunctionSpace__maxloclev_arr_int(const NodeColumns* This, const
         maximum = new int[size];
         glb_idx = new long[size];
         level   = new int[size];
-        for( size_t j=0; j<size; ++j ) {
+        for( size_t j=0; j<(size_t)size; ++j ) {
           maximum[j] = maxvec[j];
           glb_idx[j] = gidxvec[j];
           level[j] = levvec[j];
